// validation/frontend/src/app/models/auth.models.ts
export type Role = 'JOB_SEEKER' | 'EMPLOYER';
export type JobStatus = 'OPEN' | 'CLOSED' | 'FILLED';
export type ApplicationStatus = 'APPLIED' | 'UNDER_REVIEW' | 'SHORTLISTED' | 'REJECTED' | 'WITHDRAWN';
export type EmploymentStatus = 'FRESHER' | 'EMPLOYED' | 'UNEMPLOYED';

export interface RegisterRequest {
  username: string;
  fullName: string;
  email: string;
  password: string;
  confirmPassword: string;
  mobileNumber: string;
  securityQuestion: string;
  securityAnswer: string;
  location: string;
  employmentStatus: EmploymentStatus;
  role: Role;
  companyName?: string;
  industry?: string;
  companySize?: string;
  companyDescription?: string;
  website?: string;
  companyLocation?: string;
}